Moved TakaoRole
Add JWT versioned
Renamed TakaoJWT configuration on server
