+++ docs/.example.properties
@@ -21,3 +21,4 @@ jwt.takao.audience=http://0.0.0.0:8080/audience
 jwt.takao.realm=TTD
 jwt.takao.access.lifetime=
 jwt.takao.refresh.lifetime=
+jwt.takao.version=1
