+++ server/src/main/kotlin/com/takaotech/dashboard/route/administration/data/UserRoleTable.kt
@@ -5,7 +5,7 @@ import com.takaotech.dashboard.route.administration.data.user.UserTable
 import org.jetbrains.exposed.sql.Table
 
 object UserRoleTable : Table() {
-    val user = reference("user", UserTable)
+    val user = reference("kotlin_user", UserTable)
     val role = reference("role", RoleTable)
 
     override val primaryKey = PrimaryKey(user, role, name = "PK_USER_ROLE")
