+++ server/src/main/kotlin/com/takaotech/dashboard/route/administration/data/role/RoleEntity.kt
@@ -1,5 +1,6 @@
 package com.takaotech.dashboard.route.administration.data.role
 
+import com.takaotech.dashboard.model.role.TakaoRole
 import org.jetbrains.exposed.dao.Entity
 import org.jetbrains.exposed.dao.EntityClass
 import org.jetbrains.exposed.dao.id.EntityID
