--- server/src/main/kotlin/com/takaotech/dashboard/route/administration/data/role/TakaoRole.kt
@@ -1,6 +1,0 @@
-package com.takaotech.dashboard.route.administration.data.role
-
-enum class TakaoRole {
-    ADMINISTRATOR,
-    USER
-}
