+++ shared/src/commonMain/kotlin/com/takaotech/dashboard/model/role/TakaoRole.kt
@@ -1,0 +1,6 @@
+package com.takaotech.dashboard.model.role
+
+enum class TakaoRole {
+    ADMINISTRATOR,
+    BASE_USER
+}
